incremental redisplay: compute-difference-set: map over all children
compute-difference-set had suposedly an optimization that it ignored all
records outside of the pane's visible region.

1. this optimization was incorrect, because "invisible" record changes still
may be relevant to the parent (i.e advancing the cursor position)

2. it was disabled either way until now because of a wrong argument order
   (let ((everywhere (or +everywhere+ pane-viewport-reion))) ...)

The relevant parts of code are removed and adjusted accordingly. Perhaps some
clever incarnation of this optimization will be implemented in the future.