incremental redisplay: handle output-record-fixed-position
Previously this option was ignored. Right now, when the output record has a
fixed position then it is not moved when it doesn't need an update.

A good example of a useful case for this option is when some kind of a control
that causes redisplay is displayed in (updating-values (stream :cache-value t
:fixed-position t) …). We programmer doesn't want this control to move i.e
when dragging it.